% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_UI.R
\name{clustering_sekened_crosstalk}
\alias{clustering_sekened_crosstalk}
\title{Saldae Clustering CrossTalk}
\usage{
clustering_sekened_crosstalk(
  cluster_mapping = NULL,
  mds_CT = NULL,
  tisefka = NULL
)
}
\arguments{
\item{mds_CT}{MDS crosstalk ShareData object}

\item{tisefka}{clustering results}
}
\value{
t.b.d
}
\description{
UI crosstalk view
}
\author{
Farid Azouaou
}
