% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_UI.R
\name{clustering_tisefka_mds}
\alias{clustering_tisefka_mds}
\title{Saldae Clustering MDS 2D representation}
\usage{
clustering_tisefka_mds(tsclust_results = NULL)
}
\arguments{
\item{tsclust_results}{clustering results}
}
\value{
t.b.d
}
\description{
Kruskal's Non-metric Multidimensional Scaling
}
\author{
Farid Azouaou
}
