% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet_model_v01.R
\name{prophet_fit}
\alias{prophet_fit}
\title{Saldae Engine using prophet package.}
\usage{
prophet_fit(tisefka = NULL, target_variable = NULL, sald_holidays = NULL)
}
\arguments{
\item{tisefka}{training data (data frame with time variable as rownames)}

\item{target_variable}{target variable}

\item{sald_holidays}{information regarding holdays(based on country)}
}
\value{
fitted object from Prophet
}
\description{
Saldae Engine using prophet package.
}
\author{
Farid Azouaou
}
