% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaldaeClustering.R
\name{SA_clustering_core_mod}
\alias{SA_clustering_core_mod}
\title{Saldae Clustering Time series SERVER module}
\usage{
SA_clustering_core_mod(input, output, session, tisefka)
}
\arguments{
\item{tisefka}{data containing a set time serie}
}
\value{
a list containing information about clustering results
}
\description{
SERVER module for time series clustering
}
\author{
Farid Azouaou
}
