% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaldaeClustering.R
\name{SA_clustering_core_ui}
\alias{SA_clustering_core_ui}
\title{Saldae Clustering Time series UI module}
\usage{
SA_clustering_core_ui(id)
}
\arguments{
\item{id}{session ID}
}
\value{
t.b.d
}
\description{
UI module for time series clustering
}
\author{
Farid Azouaou
}
