% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_UI.R
\name{SA_clustering_mod}
\alias{SA_clustering_mod}
\title{Saldae Clustering Time series SERVER module}
\usage{
SA_clustering_mod(input, output, session, clust_results)
}
\arguments{
\item{clust_results}{list containing clustering results}
}
\value{
a list containing information about clustering results
}
\description{
SERVER module for time series clustering
}
\author{
Farid Azouaou
}
