% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaldaeClustering.R
\name{Saldae_cluserting_main}
\alias{Saldae_cluserting_main}
\title{Saldae Clustering Time series}
\usage{
Saldae_cluserting_main(
  tisefka = NULL,
  anomaly_detection = FALSE,
  num_clusters = NULL,
  clust_distance = "dtw_basic",
  clust_algo = "partitional"
)
}
\arguments{
\item{tisefka}{data containing a set time serie}
}
\value{
a list containing information about clustering results
}
\description{
proceed to
}
\author{
Farid Azouaou
}
