% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Saldae_forecasting_pool.R
\name{Saldae_Forecaster}
\alias{Saldae_Forecaster}
\title{Saldae Forecasting tool}
\usage{
Saldae_Forecaster(
  tisefka = NULL,
  target_variables = NULL,
  anomaly_detection = FALSE,
  Saldae_model = c("saldae_prophet")
)
}
\arguments{
\item{target_variables}{target variables}
}
\value{
a list of forecasts
}
\description{
Time Series Forecasting engine ....
}
\author{
Farid Azouaou
}
