% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet_model_v01.R
\name{Saldae_prophet}
\alias{Saldae_prophet}
\title{Saldae Engine(main) using prophet package.}
\usage{
Saldae_prophet(tisefka = NULL, sald_holidays = NULL, asurif_zdat = 10)
}
\arguments{
\item{tisefka}{training data (data frame with time variable as rownames)}

\item{sald_holidays}{information regarding holdays(based on country)}

\item{asurif_zdat}{steps ahead (forecast horizon)}

\item{target_variable}{target variable}
}
\value{
forecast object.
}
\description{
Saldae Engine(main) using prophet package.
}
\author{
Farid Azouaou
}
