% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcast_summary_module.R
\name{sekned_forecast_aqeru}
\alias{sekned_forecast_aqeru}
\title{Saldae display forecast results}
\usage{
sekned_forecast_aqeru(
  fcast_df = NULL,
  target_variable = NULL,
  background_colour = NULL,
  plot_settings = NULL
)
}
\arguments{
\item{fcast_df}{forecast results in tibble object}

\item{target_variable}{target variable}

\item{background_colour}{background color}
}
\value{
tibble object containing actuals , corrected, forecast and prediction bands(upper and lower)
}
\description{
t.b.d
}
\author{
Farid Azouaou
}
